Improve builds in IntelliJ and on Windows
- ScalaFile is moved to the separate, common module since in IntelliJ
the main scope can't depend on the test scope.
- Adopters.scala and Readme.scala moved to src/main so that the build
works in IntelliJ. In cmd it works as previously, unless I overlooked
something.
- The tarball with test sources is now downloaded directly in the
Scala code instead of using wget and tar. Therefore, now it works also
on Windows.

In general currently one can run compile in cmd (to generate sources)
and simply import the project to IntelliJ. At the beginning IDE will
inform that its output dirs for tests and main scope need to be split
but, once user confirmed the automatic fix, the build should work in
IntelliJ out of the box.

When one uses autocrlf = input, builds work also on Windows, but
I haven't fixed yet tests to have them all working on Windows.
