+++ bin/scalafmt.bat
@@ -1,0 +1,1 @@
+java -jar PATH_TO/scalafmt.jar %*
