+++ readme/src/main/scala/org/scalafmt/readme/Adopters.scala
